+++ src/App.jsx
@@ -1,26 +1,6 @@
-import { useRef } from "react";
+import { useRef, useState } from "react";
+import { getCalls, addToCalls } from "./firebase/firebase";
 import "./App.css";
-import { initializeApp } from "firebase/app";
-import {
-  getFirestore,
-  collection,
-  getDocs,
-  getDoc,
-} from "firebase/firestore/lite";
-
-const firebaseConfig = {
-  apiKey: "AIzaSyBXRA_-eQVL7pj8aEQXF8VQEoTxqI8qEWg",
-  authDomain: "testp2p-39de4.firebaseapp.com",
-  databaseURL: "https://testp2p-39de4-default-rtdb.firebaseio.com",
-  projectId: "testp2p-39de4",
-  storageBucket: "testp2p-39de4.appspot.com",
-  messagingSenderId: "1035623282363",
-  appId: "1:1035623282363:web:ac210588b6ee22ebe81b9e",
-};
-
-const firebaseApp = initializeApp(firebaseConfig);
-
-const db = getFirestore(firebaseApp);
 
 const servers = {
   iceServers: [
@@ -39,6 +19,7 @@ let localStream = null;
 let remoteStream = null;
 
 function App() {
+  const [callID, setCallID] = useState();
   const localVideoRef = useRef();
   const remoteVideoRef = useRef();
 
@@ -72,61 +53,24 @@ function App() {
   };
 
   const handleCreateCall = async () => {
-    const callCollectionRef = collection(db, "calls");
+    //Get candidates for caller, save to db
+    // peerConnection.onicecandidate = (event) => {
+    //   event.candidate && offerCandidates.
+    // }
 
-    // const callCollection = await getDocs(collection(db, "calls"));
-    const callDocsSnapshot = await getDocs(callCollectionRef);
+    setCallID();
 
-    let offerCandidatesCollection = null;
-
-    let answerCandidatesCollection = null;
-
-    callDocsSnapshot.forEach((callDoc) => {
-      // Aquí puedes acceder a los datos del documento principal usando callDoc.data()
-      const callId = callDoc.id;
-
-      // Para acceder a las subcolecciones, puedes seguir un enfoque similar:
-      offerCandidatesCollection = collection(callDoc.ref, "offerCandidates");
-      answerCandidatesCollection = collection(callDoc.ref, "answerCandidates");
-
-      // A partir de aquí, puedes realizar operaciones en las subcolecciones (por ejemplo, agregar documentos).
-    });
-
-    // Get candidates for caller, save to db
-    peerConnection.onicecandidate = (event) => {
-      event.candidate &&
-        offerCandidatesCollection.add(event.candidate.toJSON());
-    };
-
-    // Create offer
+    //creamos la offer del caller
     const offerDescription = await peerConnection.createOffer();
-    await peerConnection.setLocalDescription(offerDescription);
+    await peerConnection.setLocalDescription(offerDescription); //aca va el SDP
 
     const offer = {
       sdp: offerDescription.sdp,
       type: offerDescription.type,
     };
 
-    await callCollection.doc("SOME_ID").set({ offer });
-
-    // Listen for remote answer
-    callCollectionRef.doc("SOME_ID").onSnapshot((snapshot) => {
-      const data = snapshot.data();
-      if (!peerConnection.currentRemoteDescription && data?.answer) {
-        const answerDescription = new RTCSessionDescription(data.answer);
-        peerConnection.setRemoteDescription(answerDescription);
-      }
-    });
-
-    // Listen for remote ICE candidates
-    answerCandidatesCollection.onSnapshot((snapshot) => {
-      snapshot.docChanges().forEach((change) => {
-        if (change.type === "added") {
-          const candidate = new RTCIceCandidate(change.doc.data());
-          peerConnection.addIceCandidate(candidate);
-        }
-      });
-    });
+    // getCalls();
+    addToCalls(offer);
   };
 
   return (
@@ -152,7 +96,7 @@ function App() {
       <h2>3. Join a Call</h2>
       <p>Answer the call from a different browser window or device</p>
 
-      <input id="callInput" />
+      <h2>Call ID: {callID}</h2>
       <button id="answerButton" disabled>
         AnswerCLASS
       </button>
